refactor(frontend): remove browser report engine
