+++ src/frontend/src/lib/portal/use-org-scope.ts
@@ -9,11 +9,6 @@ import {
   type RosterEntry,
 } from "@/lib/insight/identity-tree";
 import {
-  reportPeopleInFlatScope,
-  reportPeopleInScope,
-  type ReportPerson,
-} from "@/lib/identities/report-person";
-import {
   type OrgScope,
 } from "@/lib/portal/portal-store";
 import {
@@ -41,7 +36,6 @@ export interface ResolvedScope {
   pivot: IdentityPerson | null;
   /** Everyone inside the scope (subtree or direct reports). */
   roster: RosterEntry[] | null;
-  reportPeople: ReportPerson[] | null;
   label: string;
   count: number;
   /** All manager nodes of the viewer's tree, for the ScopeSelect picker. */
@@ -68,7 +62,6 @@ export function flatOrgScope(
     return {
       pivot: null,
       roster: null,
-      reportPeople: null,
       label: WHOLE_ORG_LABEL,
       count: 0,
       managerNodes: [],
@@ -89,7 +82,6 @@ export function flatOrgScope(
   return {
     pivot: null,
     roster: members,
-    reportPeople: reportPeopleInFlatScope(roster),
     label: WHOLE_ORG_LABEL,
     count: members.length,
     managerNodes: [],
@@ -111,7 +103,6 @@ export function resolveScopeRoster(
     return {
       pivot: null,
       roster: null,
-      reportPeople: null,
       label: "",
       count: 0,
       managerNodes: [],
@@ -156,7 +147,6 @@ export function resolveScopeRoster(
   return {
     pivot,
     roster,
-    reportPeople: reportPeopleInScope(pivot, roster),
     label: personDisplayName(pivot),
     count: roster?.length ?? 0,
     managerNodes,
